+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `fleetman config server remove` — removes a server from the fleet
+
 ## [0.0.4] - 2026-04-16
 
 ### Added